pyproject.toml
[tool.poetry]

name = "household-services"
version = "0.1.0"
description = ""
authors = ["Vagadeeshwar G <21f1005494@ds.study.iitm.ac.in>"]
readme = "README.md"
# package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.3"
python-dotenv = "^1.0.1"
flask-sqlalchemy = "^3.1.1"
flask-restful = "^0.3.10"
flask-cors = "^5.0.0"
flask-marshmallow = "^1.2.1"
pyjwt = "^2.10.0"
marshmallow-sqlalchemy = "^1.1.0"
redis = "^5.2.0"
flask-caching = "^2.3.0"
celery = "^5.4.0"
flower = "^2.0.1"
flask-mail = "^0.10.0"

[tool.poetry.group.dev.dependencies]
faker = "^33.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
